rename: about-main-edit component
