+++ src/components/about/index.ts
@@ -1,4 +1,4 @@
 import { About } from './about'
-import { AboutMainEdit } from './about-main-edit'
+import { AboutMainEdit } from './about-main-admin'
 
 export { About, AboutMainEdit }
